pyproject.toml
[project]
name = "mlx-sam3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "daimension-python-sdk",
    "ftfy>=6.3.1",
    "huggingface-hub>=1.1.6",
    "iopath>=0.1.10",
    "ipykernel>=7.1.0",
    "ipympl>=0.9.3",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "mlx>=0.30.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sam3", "sam3.model", "rawscan_tracking", "rawscan_tracking.backends"]

[tool.uv.sources]
daimension-python-sdk = { path = "../dAImension-Python-SDK", editable = true }
